Refactor: restructure into src/yt_dont_recommend package layout
Split yt_dont_recommend.py into a proper src/ package:
  src/yt_dont_recommend/__init__.py  — all testable logic (state, blocklist,
                                       scheduler, notifications, version check, main)
  src/yt_dont_recommend/config.py    — constants and file paths (no package imports)
  src/yt_dont_recommend/browser.py   — Playwright browser automation; uses late
                                       imports from the package to avoid circular deps

Updated tests/conftest.py to add src/ to sys.path (replaces root conftest.py).
Updated pyproject.toml: packages = ["src/yt_dont_recommend"].

The test suite's monkeypatching pattern (patch("yt_dont_recommend.X")) required
all patchable-at-test-time names to reside in __init__.py rather than sub-modules;
browser.py uses _pkg() late imports for write_attention/save_state/check_removals.

Zero behaviour changes. All 109 tests pass.
